Add information on modules of features or samples

getModules and addModules generate a logical modules table from a list of of named vectors specifying which features or samples belong to each module.

getModules(x, modules, by = 1L, group = "taxonomy", exact.tax.level = FALSE)

addModules(x, modules, by = 1L, group = "taxonomy", exact.tax.level = FALSE)

# S4 method for class 'SummarizedExperiment'
addModules(x, modules, by = 1L, group = "taxonomy", exact.tax.level = FALSE)

# S4 method for class 'SummarizedExperiment'
getModules(x, modules, by = 1L, group = "taxonomy", exact.tax.level = FALSE)

Arguments

x

a SummarizedExperiment.

modules

Named character list. Named list of vectors, where each vector is a module and its elements are the module members.

by

Numeric scalar or Character scalar. Determines if modules are added row-wise (c(1L, "features")) or column-wise. (C(2L, "samples")). (Default: 1L).

group

Character vector. One or more variable names from rowData or colData, depending on by. If set to "taxonomy", modules are derived from the combined taxonomic ranks. (Default: "taxonomy")

exact.tax.level

Logical scalar. Should only the deepest taxonomic rank be used to determine whether a feature belongs to a module? if FALSE, all ranks are considered. (Default: FALSE).

Value

getModules returns a matrix where rows and columns represent features/samples and modules, respectively. addModules returns an updated SummarizedExperiment object with the modules table in the rowData or colData.

Details

getModules and addModules allow to integrate information on microbial modules into rowData(x) from databases like BugSigDB or custom tables of microbes like butyrate. sigs must be a single character vector or a list of character vectors, where each vector corresponds to a module. Each microbial signature should follow the metaphlan taxonomy format, e.g., "k__Bacteria|p__Actinobacteria|c__Actinomycetia|o__Corynebacteriales".

Examples

# Load butyrate module
data("butyrate", package = "ariadne")

# Load dataset
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Convert butyrate table to module list
butyrate <- getFullTaxonomyLabels(butyrate)

# Generate modules table
mod.table <- getModules(tse, butyrate)

# Generate & store modules table
tse <- addModules(tse, butyrate)

# Make example list of two modules
modules <- list(A = c("g__Bacteroides", "f__Enterobacteriaceae"),
                B = c("c__Clostridia", "p__Cyanobacteria"))

# Match modules with only the deepest taxonomic rank of tse
mod.table <- getModules(tse, modules, exact.tax.level = TRUE)

# Make example list of three column modules
col.modules <- list(A = c("Cohort_1", "Cohort_3"),
                    B = c("Cohort_2"),
                    c("Cohort_1", "Cohort_2", "Cohort_3"))

# Find column modules based on cohort variable
mod.table <- getModules(tse, col.modules, by = 2L, group = "cohort")