Import mappings from a file or a database

importMapping retrieves mapping information from a file or a database.

importMapping(map.file, from = NULL, to = NULL, merge = TRUE, verbose = TRUE)

# S4 method for class 'character'
importMapping(map.file, from = NA, to = NA, merge = TRUE, verbose = TRUE)

Arguments

map.file

Character vector. One or more paths to custom mapping files or one or more names from the available databases (c("ChocoPhlAn", "Woltka")).

from

Character vector. One or more strings specifying the keys from which mapping is performed. (Default: NULL)

to

Character vector. One or more strings specifying the values to which mapping is performed. (Default: NULL)

merge

Logical scalar. Should multiple mapping files be merged. (Default: TRUE).

verbose

Logical scalar. Should information on execution be printed in the console. (Default: TRUE).

Value

importMapping returns a named list of vectors, where each vector is a mapping key and its elements are the mapped values.

Details

Structure of map.file

Currently, the following databases are available:

"ChocoPhlAn": x-to-uniref mapping files from bioBakery

repo: https://zenodo.org/records/17100034

from: c("eggnog", "go", "ko", "level4ec")

to: c("uniref50", "uniref90")

"Woltka": x-to-uniref mapping files from the Web of Life

repo: https://ftp.microbio.me/pub/wol-20April2021/

from: c("eggnog", "go", "ko", "orthodb", "refseq")

to: "uniref90"

Examples

# Import eggnog-to-uniref90 mapping from ChocoPhlAn
map1 <- importMapping("ChocoPhlAn", from = "eggnog", to = "uniref90")
#> Retrieving mappings from /github/home/.cache/R/ariadne/21385bd32039_map_eggnog_uniref90.txt.gz.

# Import ko-to-uniref90 mapping from ChocoPhlAn
map2 <- importMapping("ChocoPhlAn", from = "ko", to = "uniref90")
#> Retrieving mappings from /github/home/.cache/R/ariadne/213819235d6a_map_ko_uniref90.txt.gz.

# Import and merge ko- and eggnog-to-uniref90 mappings from ChocoPhlAn
map3 <- importMapping(
    "ChocoPhlAn",
    from = c("eggnog", "ko"),
    to = "uniref90"
)
#> Retrieving mappings from /github/home/.cache/R/ariadne/21385bd32039_map_eggnog_uniref90.txt.gz.
#> Retrieving mappings from /github/home/.cache/R/ariadne/213819235d6a_map_ko_uniref90.txt.gz.

# Import local mapping file
# map4 <- importMapping("path/to/file")