List resource versions registered in the ariadne database

listResourceVersions shows the available versions for the different resources registered in the ariadne database. Any of those versions can be passed to ariadne. Resources labelled with "latest" cannot be versioned as they are dynamically accessed via their API or SPARQL endpoint.

listResourceVersions(default = FALSE)

Arguments

default

Logical scalar. Should only default versions be listed. (Default: FALSE)

Value

A data.frame with information on registered resource versions.

See also

Examples

# View all available resource versions
listResourceVersions()
#>      resource      version
#> 1    BugSigDB       v1.3.0
#> 2    BugSigDB       v1.2.2
#> 3    BugSigDB       v1.2.1
#> 4    BugSigDB       v1.2.0
#> 5    BugSigDB       v1.1.0
#> 6    BugSigDB       v1.0.2
#> 7    BugSigDB       v1.0.1
#> 8    BugSigDB       v1.0.0
#> 9  ChocoPhlAn     v201901b
#> 15         GM           v1
#> 10         GO   2026-03-25
#> 11         GO   2026-01-23
#> 12         GO   2025-10-10
#> 17       KEGG       latest
#> 18        OTT       latest
#> 19       Rhea       latest
#> 16   TIGRFAMs          v15
#> 20    UniProt       latest
#> 13        WoL           v2
#> 14        WoL v20April2021
#>                                                                        url
#> 1                                     https://zenodo.org/records/15272273/
#> 2                                     https://zenodo.org/records/13997429/
#> 3                                     https://zenodo.org/records/10627578/
#> 4                                     https://zenodo.org/records/10407666/
#> 5                                      https://zenodo.org/records/6468009/
#> 6                                      https://zenodo.org/records/5904281/
#> 7                                      https://zenodo.org/records/5819260/
#> 8                                      https://zenodo.org/records/5606166/
#> 9                                     https://zenodo.org/records/17100034/
#> 15 https://github.com/omixer/omixer-rpmR/raw/refs/heads/main/inst/extdata/
#> 10                            https://release.geneontology.org/2026-03-25/
#> 11                            https://release.geneontology.org/2026-01-23/
#> 12                            https://release.geneontology.org/2025-10-10/
#> 17                                             https://www.genome.jp/kegg/
#> 18                                       https://opentreeoflife.github.io/
#> 19                                                https://www.rhea-db.org/
#> 16                 https://ftp.ncbi.nlm.nih.gov/hmm/TIGRFAMs/release_15.0/
#> 20                                                https://www.uniprot.org/
#> 13                                       https://ftp.microbio.me/pub/wol2/
#> 14                            https://ftp.microbio.me/pub/wol-20April2021/

# View default resource versions
listResourceVersions(default = TRUE)
#>      resource    version
#> 1    BugSigDB     v1.3.0
#> 9  ChocoPhlAn   v201901b
#> 15         GM         v1
#> 10         GO 2026-03-25
#> 17       KEGG     latest
#> 18        OTT     latest
#> 19       Rhea     latest
#> 16   TIGRFAMs        v15
#> 20    UniProt     latest
#> 13        WoL         v2
#>                                                                        url
#> 1                                     https://zenodo.org/records/15272273/
#> 9                                     https://zenodo.org/records/17100034/
#> 15 https://github.com/omixer/omixer-rpmR/raw/refs/heads/main/inst/extdata/
#> 10                            https://release.geneontology.org/2026-03-25/
#> 17                                             https://www.genome.jp/kegg/
#> 18                                       https://opentreeoflife.github.io/
#> 19                                                https://www.rhea-db.org/
#> 16                 https://ftp.ncbi.nlm.nih.gov/hmm/TIGRFAMs/release_15.0/
#> 20                                                https://www.uniprot.org/
#> 13                                       https://ftp.microbio.me/pub/wol2/