Map modules to taxa

mapModules returns a list of modules containing the taxa that are members of each module. Taxa are derived from uniref ids by querying UniProt SPARQL. Membership is based on whether a taxon meets the criteria specified by a module.

mapModules(modules, map, remove.empty = TRUE, message = TRUE)

# S4 method for class 'list'
mapModules(modules, map, remove.empty = TRUE, message = TRUE)

Arguments

modules

Named character list. Named list of vectors, where each vector is a module and its elements are the module members.

map

Named character list. Named list of vectors, where each vector is a mapping key and its elements are the mapped values.

remove.empty

Logical scalar. Should modules with no matching taxa be removed. (Default: TRUE).

message

Logical scalar. Should information on execution be printed in the console. (Default: TRUE).

x

a TreeSummarizedExperiment.

Value

mapModules returns a named list of vectors, where each vector is a module and its elements are the taxa members of that module.

Details

Additional details

Examples

# Import GBM
gbm <- importModules("GBM")
#> adding rname 'GBMs.v1.0.txt'
#> Warning: download failed
#>   web resource path: ‘GBMs.v1.0.txt’
#>   local file path: ‘/home/giulio/.cache/R/ariadne/45157199ead7_GBMs.v1.0.txt’
#>   reason: Couldn't resolve host name [GBMs.v1.0.txt]:
#> Could not resolve host: GBMs.v1.0.txt
#> Warning: bfcadd() failed; resource removed
#>   rid: BFC8
#>   fpath: ‘GBMs.v1.0.txt’
#>   reason: download failed
#> Warning: 
#> trying to add rname 'GBMs.v1.0.txt' produced error:
#>   bfcadd() failed; see warnings()
#> Error in bfcrpath(bfc, rnames = url, exact = TRUE, download = TRUE, rtype = "web"): not all 'rnames' found or unique.
# Import ko-to-uniref90 mapping
map1 <- importMapping("ko.uniref90")
#> Retrieving mappings from ko.uniref90.
#> Warning: cannot open file 'ko.uniref90': No such file or directory
#> Error in file(con, "r"): cannot open the connection
# Map modules to taxa
sigs <- mapModules(gbm[seq(3)], map1)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'modules' in selecting a method for function 'mapModules': object 'gbm' not found