Weave paths between resources

weavePath and weaveComplex bridge the path between resources, fetching and combining the necessary data from the ariadne database.

weavePath(graph, ...)

weaveComplex(graph, ...)

# S4 method for class 'igraph'
weavePath(
  graph,
  by,
  k = 1,
  include = NULL,
  exclude = NULL,
  init = NULL,
  prune = TRUE,
  use.names = TRUE,
  verbose = TRUE,
  timeout = 1e+06,
  ...
)

# S4 method for class 'igraph'
weaveComplex(
  graph,
  by,
  k = 1,
  include = NULL,
  exclude = NULL,
  init = NULL,
  prune = TRUE,
  use.names = TRUE,
  mode = "presence",
  threshold = 1,
  verbose = TRUE,
  timeout = 1e+06,
  ...
)

Arguments

graph

An igraph object.

...

Additional arguments.

batch.size: Numeric scalar. The maximum batch size for SPARQL or API queries. (Default: half the maximum query size)

workers: Numeric scalar. Number of workers to use, automatically detected when NULL. (Default: NULL)

factor: Numeric scalar. Number of jobs per worker. (Default: 3)

by

A formula specifying the path to weave.

k

Numeric scalar. The kth shortest path to weave. (Default: 1)

include

Character vector. Nodes to cross in the path. (Default: NULL)

exclude

Character vector. Nodes to avoid in the path. (Default: NULL)

init

Character vector. Initial values to prune the first mapping step. (Default: NULL)

prune

Logical scalar. Should the values from each step be used prune the following step. (Default: TRUE)

use.names

Logical scalar. Should feature names be used in the output instead of feature identifiers. either (Default: TRUE)

verbose

Logical scalar. Should messages be printed in the console. (Default: TRUE)

timeout

Numeric scalar. The timeout for downloading resources. (Default: 1e6)

mode

Character scalar. The mode of the output, either as "presence" or "coverage" information. Only for weaveComplex. (Default: "presence")

threshold

Numeric scalar. The coverage threshold to infer presence, between 0 and 1. Only for weaveComplex. (Default: 1)

Value

A two-column data.frame (x-to-y linkmap).

Examples

library(mia)
#> Loading required package: MultiAssayExperiment
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: generics
#> 
#> Attaching package: ‘generics’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: Seqinfo
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> Loading required package: SingleCellExperiment
#> Loading required package: TreeSummarizedExperiment
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> This is mia version 1.19.7
#> - Online documentation and vignettes: https://microbiome.github.io/mia/
#> - Online book 'Orchestrating Microbiome Analysis (OMA)': https://microbiome.github.io/OMA/docs/devel/

# Import dataset
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Load resource graph
graph <- ariadne()

# Retrieve taxon names
tax.labs <- getTaxonomyLabels(tse, make.unique = FALSE)
tax.labs <- sub("^.+:", "", tax.labs)

# Weave path starting from initial values
tax2bugsig <- weavePath(graph, taxname ~ bugsig, init = tax.labs)
#> taxname -(BugSigDB)-> bugsig

# Weave 3-rd path
tax2bugsig <- weavePath(graph, taxname ~ bugsig, init = tax.labs, k = 3)
#> taxname -(OTT)-> taxid
#> taxid -(BugSigDB)-> bugsig

# Weave path including taxid
tax2bugsig <- weavePath(
    graph, taxname ~ bugsig, include = "taxid", init = tax.labs
)
#> taxname -(UniProt)-> taxid
#> taxid -(BugSigDB)-> bugsig

# Weave simple path from KEGG diseases to gut metabolic modules
dis2gmm <- weavePath(graph, disease ~ gmm)
#> disease -(KEGG)-> ko
#> ko -(GM)-> gmm

# Weave complex path from KEGG diseases to gut metabolic modules
dis2gmm <- weaveComplex(graph, disease ~ gmm, threshold = 0.8)
#> disease -(KEGG)-> ko
#> ko -(GM)-> gmm

# Obtain results in terms of coverage
dis2gmm <- weaveComplex(graph, disease ~ gmm, mode = "coverage")
#> disease -(KEGG)-> ko
#> ko -(GM)-> gmm